% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varSelRanger.R
\name{varSelRanger}
\alias{varSelRanger}
\title{Variable selection using recursive feature elimination.}
\usage{
varSelRanger(data, dependent.variable.name, formula = NULL, frac = 0.2,
  min.vars = 1, importance = "impurity", ...)
}
\arguments{
\item{data}{Training data of class \code{data.frame}, \code{matrix} or \code{dgCMatrix}.}

\item{dependent.variable.name}{Name of the dependent variable. For survival forests this is the time variable.}

\item{formula}{Object of class formula or character describing the model to fit. Interaction terms supported only for numerical variables. (Formula interface not yet supported in varSelRanger).}

\item{frac}{Fraction of top variables to drop after each iteration.}

\item{min.vars}{Minimum number of variables allowed in a model.}

\item{importance}{Variable importance mode. One of 'none', 'impurity', 'impurity_corrected', 'permutation'. The 'impurity' measure is the Gini index for classification, the variance of the responses for regression and the sum of test statistics (see splitrule) for survival.}

\item{...}{Further arguments to be passed to \link{\code{ranger::ranger}}.}
}
\description{
Models are evaluated based on the OOB error rate.
}
